Extract mozfun docs generation and modify call order
